# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Spacebox API",
    "description": "for spacebox",
    "version": "1.0"
  },
  "produces": ["application/json", "text/plain"],
  "host": "spacebox-tech.herokuapp.com",
  "basePath": "/",
  "definitions": {
...
    "uuid": {
      "type": "string",
      "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
    },
    "responseError": {
      "type": "object",
      "properties": {
        "errorDetails": {
          "description": "this really should be rationalized but atm, it could be a string OR an object. Swagger doesn't support this as a schema so it's left undefined"
        }
      }
    },
...
  },
  "securityDefinitions": {
      "swagger-ui-key": {
        "description": "This just defines what the Swagger-UI does so it has to be supported. The preferred method is to pass in the JWT token as a bearer token in the Authorization header.",
        "name": "api_key",
        "type": "apiKey",
        "in": "query"
      }
  },
  "security": [
    {
      "swagger-ui-key": []
    }
  ],
  "responses": {
    "Error": {
      "description": "Any kind of error, not just an internal error",
      "schema": {
        "$ref": "#/definitions/responseError"
      }
    },
    "Empty": {
      "description": "It's a normal empty response, but swagger can't understand that",
      "schema": {
        "type": "string",
        "enum": [ "" ]
      }
    },
    "NotFound": {
      "description": "Couldn't find something you specified",
      "schema": {
        "type": "string"
      }
    }
  },
  "paths": {
...
    "/jobs/{uuid}": {
      "post": {
        "tags": [ "jobs" ],
        "description": "Currently the only effect is that the job will be retriggered for processing. When a job fails the backoff is increased and if you fix the error condition, it may be awhile before the job is checked again. Call this endpoint and the job will be checked soon.",
        "parameters": [
          {
            "name": "uuid",
            "description": "the uuid of the job",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "204": {"$ref": "#/responses/Empty" }
        }
      },
      "delete": {
        "tags": [ "jobs" ],
        "description": "If your jobs errors out, you'll need to resolve the error or delete the job before the facility will start processing jobs again. If you delete a job that has already consumed the resources, you won't get them back.",
        "parameters": [
          {
            "name": "uuid",
            "description": "the uuid of the job",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "204": {"$ref": "#/responses/Empty" }
        }
      },
      "get": {
        "tags": [ "jobs" ],
        "description": "returns the details of the job",
        "parameters": [
          {
            "name": "uuid",
            "description": "the uuid of the job",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "304": { "$ref": "#/responses/Empty" },
          "404": { "$ref": "#/responses/NotFound" },
          "200": {
            "description": "A job",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/facilities/{uuid}": {
      "delete": {
        "tags": [ "facilities" ],
        "description": "delete facilities that you uninstalled during a refit or construction. You can only delete disabled facilities. Deleting a facility cancels any jobs running there. Cancelling a job you lose both the output and the input, that may change in the future.",
        "parameters": [
          {
            "name": "uuid",
            "description": "the uuid of the blueprint you want",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "204": { "$ref": "#/responses/Empty" }
        }
      }
    },
    "/facilities": {
      "post": {
        "tags": [ "facilities" ],
        "description": "re-enable any facilities that were disabled during refitting or construction",
        "parameters": [],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": {
            "description": "Facilities",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        }
      },
      "get": {
        "tags": [ "facilities" ],
        "description": "get a list of all your facilities that you can run jobs in",
        "parameters": [],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "304": { "$ref": "#/responses/Empty" },
          "200": {
            "description": "Facilities",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "get": {
        "tags": [ "jobs" ],
        "description": "get a list of all your jobs, past and present. There is no garbage collection yet",
        "parameters": [],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "304": { "$ref": "#/responses/Empty" },
          "200": {
            "description": "Jobs",
            "schema": {
              "type": "array",
              "items": {
                "type": "object"
              }
            }
          }
        }
      },
      "post": {
        "deprecated": true,
        "tags": [ "jobs" ],
        "description": "start a job",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "the parameters to the job as json",
            "schema": {
              "type": "object",
              "properties": {
                "facility": {
                  "description": "the uuid of the facility run the job in",
                  "$ref": "#/definitions/uuid"
                },
                "action": {
                  "type": "string",
                  "enum": [ "construction", "refining", "refitting", "research", "manufacturing" ]
                }
              }
            }
          }
        ],
        "responses": {
          "default": { "$ref": "#/responses/Error" },
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "job": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "description": "the uuid of the job",
                      "$ref": "#/definitions/uuid"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}